Promiseを返す命令に「保障」の語句を追加して統一。 Promiseを作成・操作する命令のpluginを既定に変更。
