+++ src/plugin_browser_ajax.js
@@ -139,6 +139,17 @@ module.exports = {
     },
     return_none: true
   },
+  'AJAX保障送信': { // @非同期通信(Ajax)でURLにデータの送信を開始する非同期処理オブジェクト(Promise)を作成する。 // @AJAXほしょうそうしん
+    type: 'func',
+    josi: [['まで', 'へ', 'に']],
+    pure: true,
+    fn: function (url, sys) {
+      let options = sys.__v0['AJAXオプション']
+      if (options === '') {options = {method: 'GET'}}
+      return fetch(url, options)
+    },
+    return_none: false
+  },
   'HTTP取得': { // @逐次実行構文にて、非同期通信(Ajax)でURLにデータを送信する。成功すると『対象』にデータが代入される。失敗すると『AJAX失敗時』を実行。 // @HTTPしゅとく
     type: 'func',
     josi: [['の', 'から', 'を']],
@@ -158,6 +169,15 @@ module.exports = {
       sys.__exec('AJAX逐次送信', [url, sys])
     },
     return_none: true
+  },
+  'HTTP保障取得': { // @非同期通信(Ajax)でURLにデータの送信を開始する非同期処理オブジェクト(Promise)を作成する。 // @HTTPほしょうしゅとく
+    type: 'func',
+    josi: [['の', 'から', 'を']],
+    pure: true,
+    fn: function (url, sys) {
+      return sys.__exec('AJAX保障送信', [url, sys])
+    },
+    return_none: false
   },
   'POST逐次送信': { // @逐次実行構文にて、AjaxでURLにPARAMSをPOST送信し『対象』にデータを設定。失敗すると『AJAX失敗時』を実行。 // @POSTちくじそうしん
     type: 'func',
@@ -197,6 +217,23 @@ module.exports = {
     },
     return_none: true
   },
+  'POST保障送信': { // @非同期通信(Ajax)でURLにPARAMSをPOST送信を開始する非同期処理オブジェクト(Promise)を作成する。 // @POSTほしょうそうしん
+    type: 'func',
+    josi: [['まで', 'へ', 'に'], ['を']],
+    pure: true,
+    fn: function (url, params, sys) {
+      let bodyData = sys.__exec('POSTデータ生成', [params, sys])
+      const options = {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/x-www-form-urlencoded'
+        },
+        body: bodyData
+      }
+      return fetch(url, options)
+    },
+    return_none: false
+  },
   'POSTフォーム逐次送信': { // @逐次実行構文にて、AjaxでURLにPARAMSをフォームとしてPOST送信し『対象』にデータを設定。失敗すると『AJAX失敗時』を実行。 // @POSTふぉーむちくじそうしん
     type: 'func',
     josi: [['まで', 'へ', 'に'], ['を']],
@@ -234,5 +271,47 @@ module.exports = {
       sys.__exec('POSTフォーム逐次送信', [url, params, sys])
     },
     return_none: true
+  },
+  'POSTフォーム保障送信': { // @非同期通信(Ajax)でURLにPARAMSをフォームとしてPOST送信を開始する非同期処理オブジェクト(Promise)を作成する。  // @POSTふぉーむほしょうそうしん
+    type: 'func',
+    josi: [['まで', 'へ', 'に'], ['を']],
+    pure: true,
+    fn: function (url, params, sys) {
+      const fd = new FormData()
+      for (let key in params)
+        {fd.set(key, params[key])}
+
+      let options = {
+        method: 'POST',
+        body: fd
+      }
+      return fetch(url, options)
+    },
+    return_none: false
+  },
+  'AJAX内容取得': { // @非同期通信(Ajax)の応答から内容を指定した形式で取り出すための非同期処理オブジェクト(Promise)を返す。  // @AJAXないようしゅとく
+    type: 'func',
+    josi: [['から'], ['で']],
+    pure: true,
+    fn: function (res, type, sys) {
+      type = type.toString().toUpperCase()
+      if (type === 'TEXT' || type === 'テキスト') {
+        return res.text()
+      } else
+      if (type === 'JSON') {
+        return res.json()
+      } else
+      if (type === 'BLOB') {
+        return res.blob()
+      } else
+      if (type === 'ARRAY' || type === '配列') {
+        return res.arrayBuffer()
+      } else
+      if (type === 'BODY' || type === '本体') {
+        return res.body
+      }
+      return res.body()
+    },
+    return_none: false
   }
 }
