+++ src/plugin_promise.js
@@ -1,39 +1,4 @@
 const PluginPromise = {
-  '初期化': {
-    type: 'func',
-    josi: [],
-    pure: true,
-    fn: function (sys) {
-      if (sys.__promise == null) {
-        sys.__promise = {
-          clearTimeoutAll: function () {
-            for (const t of this.__timeoutList) { clearTimeout(t) }
-          },
-          clearTimeoutById: function (timerId) {
-            const i = this.__timeoutList.indexOf(timerId)
-            if (i >= 0) {this.__timeoutList.splice(i, 1)}
-          },
-          addTimeoutId: function (timerId) {
-            this.__timeoutList.unshift(timerId)
-          },
-          setLastPromise: function (promise) {
-            sys.__v0['そ'] = promise
-            return promise
-          }
-        }
-      }
-      sys.__promise.__timeoutList = []
-    }
-  },
-  '!クリア': {
-    type: 'func',
-    josi: [],
-    pure: true,
-    fn: function (sys) {
-      if (sys.__promise == null) {return}
-      sys.__promise.clearTimeoutAll()
-    }
-  },
   // @非同期処理
   'そ': {type: 'const', value: ''}, // @そ
   // @非同期処理
@@ -107,116 +72,7 @@ const PluginPromise = {
       return sys.__promise.setLastPromise(Promise.all(args))
     },
     return_none: false
-  },
-
-  // @非同期処理/システム系入れ替え
-  '秒後': { // @N秒後に成功する非同期処理オブジェクト(Promise)を作成する // @びょうご
-    type: 'func',
-    josi: [['']],
-    pure: true,
-    fn: function (n, sys) {
-      return new Promise(resolve => {
-        // 1回限りのタイマーをセット
-        const timerId = setTimeout(() => {
-          // 使用中リストから完了したIDを削除
-          sys.__promise.clearTimeoutById(timerId)
-          resolve(timerId)
-        }, parseFloat(n) * 1000)
-        // 使用中リストに開始したIDを追加
-        sys.__promise.addTimeoutId(timerId)
-      })
-    },
-    return_none: false
-  },
-  // @非同期処理/ajax系入れ替え
-  'AJAX送信': { // @非同期通信(Ajax)でURLにデータの送信を開始する非同期処理オブジェクト(Promise)を作成する。 // @AJAXそうしん
-    type: 'func',
-    josi: [['まで', 'へ', 'に']],
-    pure: true,
-    fn: function (url, sys) {
-      let options = sys.__v0['AJAXオプション']
-      if (options === '') {options = {method: 'GET'}}
-      return fetch(url, options)
-    },
-    return_none: false
-  },
-  'HTTP取得': { // @非同期通信(Ajax)でURLにデータの送信を開始する非同期処理オブジェクト(Promise)を作成する。 // @HTTPしゅとく
-    type: 'func',
-    josi: [['の', 'から', 'を']],
-    pure: true,
-    fn: function (url, sys) {
-      return sys.__exec('AJAX送信', [url, sys])
-    },
-    return_none: false
-  },
-  'POST送信': { // @非同期通信(Ajax)でURLにPARAMSをPOST送信を開始する非同期処理オブジェクト(Promise)を作成する。 // @POSTそうしん
-    type: 'func',
-    josi: [['まで', 'へ', 'に'], ['を']],
-    pure: true,
-    fn: function (url, params, sys) {
-      let bodyData = sys.__exec('POSTデータ生成', [params, sys])
-      const options = {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/x-www-form-urlencoded'
-        },
-        body: bodyData
-      }
-      return fetch(url, options)
-    },
-    return_none: false
-  },
-  'POSTフォーム送信': { // @非同期通信(Ajax)でURLにPARAMSをフォームとしてPOST送信を開始する非同期処理オブジェクト(Promise)を作成する。  // @POSTふぉーむそうしん
-    type: 'func',
-    josi: [['まで', 'へ', 'に'], ['を']],
-    pure: true,
-    fn: function (url, params, sys) {
-      const fd = new FormData()
-      for (let key in params)
-        {fd.set(key, params[key])}
-
-      let options = {
-        method: 'POST',
-        body: fd
-      }
-      return fetch(url, options)
-    },
-    return_none: false
-  },
-  '内容取得': { // @非同期通信(Ajax)の応答から内容を指定した形式で取り出すための非同期処理オブジェクト(Promise)を返す。  // @内容取り出す
-    type: 'func',
-    josi: [['から'], ['で']],
-    pure: true,
-    fn: function (res, type, sys) {
-      type = type.toString().toUpperCase()
-      if (type === 'TEXT' || type === 'テキスト') {
-        return res.text()
-      } else
-      if (type === 'JSON') {
-        return res.json()
-      } else
-      if (type === 'BLOB') {
-        return res.blob()
-      } else
-      if (type === 'ARRAY' || type === '配列') {
-        return res.arrayBuffer()
-      } else
-      if (type === 'BODY' || type === '本体') {
-        return res.body
-      }
-      return res.body()
-    },
-    return_none: false
-  }  
+  }
 }
 
-if (typeof global === 'object' && typeof global.module !== 'undefiled') {
-  module.exports = PluginPromise
-}
-// scriptタグで取り込んだ時、自動で登録する
-/* istanbul ignore else */
-if (typeof (navigator) === 'object' && typeof (navigator.nako3) === 'object') 
-  {navigator.nako3.addPluginObject('PluginPromise', PluginPromise)}
-
-
-
+module.exports = PluginPromise
